Rename Chart namespace to Charting to avoid naming conflicts
